added jwt to layout

+++ api/src/layout/layout.controller.ts
@@ -13,45 +13,59 @@ import { Body,
 import { FindLayoutsDto } from './dto/find-layouts.dto';
 import { CreateLayoutDto } from './dto/create-layout.dto';
 import { LayoutService } from './layout.service';
-import { REVIEW_NOT_FOUND } from './layout.constans';
+import { LAYOUT_NOT_FOUND, LAYOUT_NOT_USER, NOT_ADMIN } from './layout.constans';
 import { transliterate } from './helpers/transliter';
 import { JwtAuthGuard } from '../auth/guards/jwt.guard';
 import { UserGuard } from '../decorators/user.decorator';
+import { Types } from 'mongoose';
 
 @Controller('layout')
 export class LayoutController {
 	constructor(private readonly layoutService: LayoutService) {
 	}
 
+	@UseGuards(JwtAuthGuard)
 	@UsePipes(new ValidationPipe())
 	@Post('create')
-	async create(@Body() dto: CreateLayoutDto) {
+	async create(@Body() dto: CreateLayoutDto, @UserGuard() guard: {_id: string}) {
 		const slash = transliterate(dto.title);
-		return this.layoutService.create({...dto, slash});
+		return this.layoutService.create({...dto, slash, user: guard._id});
 	}
 
+	@UseGuards(JwtAuthGuard)
 	@Delete(':id')
-	async delete(@Param('id') id: string) {
+	async delete(@Param('id') id: string, @UserGuard() guard: {_id: Types.ObjectId, email: string}) {
+		const layout = await this.layoutService.findById(id);
+		if(layout && layout.user !== guard._id) {
+			throw new HttpException(LAYOUT_NOT_USER, HttpStatus.BAD_REQUEST);
+		}
+
 		const deletedLayout = await this.layoutService.delete(id);
 		if(!deletedLayout) {
-			throw new HttpException(REVIEW_NOT_FOUND, HttpStatus.NOT_FOUND);
+			throw new HttpException(LAYOUT_NOT_FOUND, HttpStatus.BAD_REQUEST);
 		}
+		return deletedLayout;
 	}
 
+	@UseGuards(JwtAuthGuard)
 	@Get(':id')
-	async get(@Param('id') id: string) {
+	async get(@Param('id') id: string, @UserGuard() guard: {_id: Types.ObjectId, email: string}) {
 		const findLayout = await this.layoutService.findById(id);
 		if(!findLayout) {
-			throw new HttpException(REVIEW_NOT_FOUND, HttpStatus.NOT_FOUND);
+			throw new HttpException(LAYOUT_NOT_FOUND, HttpStatus.NOT_FOUND);
+		}
+		if(findLayout.user !== guard._id) {
+			throw new HttpException(LAYOUT_NOT_USER, HttpStatus.BAD_REQUEST);
 		}
 		return findLayout;
 	}
 
+	@UseGuards(JwtAuthGuard)
 	@Get('user/:id')
-	async findByUser(@Body() id: string) {
+	async findByUser(@Param('id') id: Types.ObjectId) {
 		const findLayouts = await this.layoutService.findByUser(id);
 		if(!findLayouts) {
-			throw new HttpException(REVIEW_NOT_FOUND, HttpStatus.NOT_FOUND);
+			throw new HttpException(LAYOUT_NOT_FOUND, HttpStatus.NOT_FOUND);
 		}
 		return findLayouts;
 	}
@@ -59,15 +73,23 @@ export class LayoutController {
 	@UseGuards(JwtAuthGuard)
 	@Get()
 	async find(@Body() dto: FindLayoutsDto, @UserGuard() guard: {email: string, _id: string, role: string}) {
+		if (guard.role !== 'admin'){
+			throw new HttpException(NOT_ADMIN, HttpStatus.BAD_REQUEST);
+		}
 		return this.layoutService.findAll(dto);
 	}
 
+	@UseGuards(JwtAuthGuard)
 	@UsePipes(new ValidationPipe())
 	@Patch(':id')
-	async patch(@Param('id') id: string, @Body() dto: CreateLayoutDto) {
+	async patch(@Param('id') id: string, @Body() dto: CreateLayoutDto, @UserGuard() guard: { _id: Types.ObjectId}) {
 		const editLayout = await this.layoutService.edit(id);
 		if(!editLayout) {
-			throw new HttpException(REVIEW_NOT_FOUND, HttpStatus.NOT_FOUND);
+			throw new HttpException(LAYOUT_NOT_FOUND, HttpStatus.NOT_FOUND);
+		}
+
+		if(editLayout.user !== guard._id) {
+			throw new HttpException(LAYOUT_NOT_USER, HttpStatus.BAD_REQUEST);
 		}
 		return editLayout;
 	}
